remove PyErr_Format arity select define

+++ graalpython/com.oracle.graal.python.cext/src/unicodeobject.c
@@ -206,7 +206,7 @@ PyObject* PyUnicode_FromFormat(const char* format, ...) {
     int argc = polyglot_get_arg_count();
     void **args = truffle_managed_malloc(sizeof(void*) * (argc - 1));
     for (int i = 1; i < argc; i++) {
-        args[i - 1] = (unsigned long)polyglot_get_arg(i);
+        args[i - 1] = polyglot_get_arg(i);
     }
     return PyTruffle_Unicode_FromFormat(format, NULL, args, 0);
 }
